requirements.txt
fastapi
uvicorn
pyyaml
git+https://github.com/YumYummity/pjsekai-background-gen-pillow-upd
asyncpg
aioboto3
authlib